[#129934961] fixed some edge cases regarding sync link

+++ sovrin/client/link_invitation.py
@@ -150,11 +150,11 @@ class LinkInvitation:
             if second_diff < 120:
                 return "a minute ago"
             if second_diff < 3600:
-                return str(second_diff / 60) + " minutes ago"
+                return str(int(second_diff / 60)) + " minutes ago"
             if second_diff < 7200:
                 return "an hour ago"
             if second_diff < 86400:
-                return str(second_diff / 3600) + " hours ago"
+                return str(int(second_diff / 3600)) + " hours ago"
         if day_diff == 1:
             return "Yesterday"
         if day_diff < 7:
